Loan Payment and Remittance Analysis Project/loan_remittance_analysis/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
faker>=18.0.0
pytest>=7.0.0
python-dateutil>=2.8.0
exchange_calendars>=4.2.0
